requirements.txt
flask
flask-cors
openai
wikipedia
python-dotenv
